Updated perform_alignment_error to include user defined alignment threshold.
Updated perform_alignment_error function in IAS to include a variable which controls the alignment error threshold. Also updated documentation that was not present for existing variables in the perform_alignment_error function. I found that there were gaps in the aligned speed signal where there was no data. The issue was that the alignment error threshold was too strict causing some parts of the signal to have no speed information as those sections were not aligned appropriately.